# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Druse.Host.LocalApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "CloseSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/connections": {
      "get": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "GetConnections",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "CreateConnection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/connections/{id}": {
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "UpdateConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "DeleteConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/connections/{id}/sessions": {
      "post": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "OpenSavedSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/OpenSavedSessionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "GetHistory",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "ClearHistory",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/preferences/{key}": {
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "SetPreference",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceValueDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workspace/tabs": {
      "get": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "GetEditorTabs",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "SaveEditorTabs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EditorTabDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/workspace/snippets/{id}": {
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "SaveSqlSnippet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SqlSnippetDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "DeleteSqlSnippet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workspace/diagrams": {
      "get": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "GetDiagrams",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workspace/diagrams/{id}": {
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "SaveDiagram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedDiagramDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "DeleteDiagram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workspace/compositions": {
      "get": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "GetCompositions",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workspace/compositions/{id}": {
      "put": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "SaveComposition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedCompositionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Druse.Host.LocalApi"
        ],
        "operationId": "DeleteComposition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionProfileDto": {
        "required": [
          "name",
          "engine",
          "host",
          "port",
          "database"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "engine": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "database": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "authentication": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "readOnly": {
            "type": "boolean"
          },
          "sslMode": {
            "type": "string"
          },
          "connectTimeoutSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sshTunnel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SshTunnelDto"
              }
            ]
          },
          "informixServer": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EditorTabDto": {
        "required": [
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isDirty": {
            "type": "boolean"
          },
          "connectionId": {
            "type": [
              "null",
              "string"
            ]
          },
          "database": {
            "type": [
              "null",
              "string"
            ]
          },
          "fileName": {
            "type": [
              "null",
              "string"
            ]
          },
          "documentId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OpenSavedSessionRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "sshSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "sshVerificationCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PreferenceValueDto": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
      "SaveConnectionRequest": {
        "required": [
          "profile"
        ],
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ConnectionProfileDto"
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "storePassword": {
            "type": "boolean"
          },
          "sshSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "storeSshSecret": {
            "type": "boolean"
          }
        }
      },
      "SavedCompositionDto": {
        "required": [
          "id",
          "connectionId",
          "database",
          "table",
          "name",
          "model"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "connectionId": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "schema": {
            "type": [
              "null",
              "string"
            ]
          },
          "table": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "SavedDiagramDto": {
        "required": [
          "id",
          "connectionId",
          "name",
          "model"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "connectionId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "SqlSnippetDto": {
        "required": [
          "id",
          "name",
          "sql"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "SshTunnelDto": {
        "required": [
          "host",
          "username"
        ],
        "type": "object",
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "authentication": {
            "type": "string"
          },
          "privateKeyPath": {
            "type": "string"
          },
          "connectTimeoutSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}